requirements.txt
aiogram==3.4.1
openai==1.70.0
python-dotenv==1.0.1
pydantic==2.5.3